[Hub] Render plugins' reference pages with their corresponding schema
