+++ app/_plugins/generators/plugins/pages/base.rb
@@ -31,7 +31,9 @@ module Jekyll
             'examples' => examples,
             'tools' => @plugin.formats,
             'breadcrumbs' => ['/plugins/'],
-            'compatible_protocols' => compatible_protocols
+            'compatible_protocols' => compatible_protocols,
+            'schema' => schema.to_json,
+            'plugin' => @plugin
           )
         end
